Updated convert.py
I created a menu for conversion to kph, m/s, and ft/s. They had to put in an input number in mph, then the menu would ask them to choose conversion type, and if within 1-3, it will convert to the choice they were given. If outside the choice option, error will appear.